+++ R/rfrailty.R
@@ -2,7 +2,7 @@
 #' @export
 #' @aliases rfrailty
 #' @description The frailty function for adding a simple random effects term to the linear predictor of a given survival regression model.
-#' @param cluster the variable to be entered as a random effect (always converted to a factor object internally).
+#' @param cluster a vector determining the grouping of subjects (always converted to a factor object internally.
 #' @param frailty the frailty distribution; current implementation includes the gamma (default), lognormal and positive stable (ps) distributions.
 #' @param sigma standard deviation assumed for the frailty distribution; sigma = 1 by default; this value is ignored for positive stable (ps) distribution.
 #' @param alpha stability parameter of the positive stable distribution; alpha must lie in (0,1) interval and an NA is return otherwise.