+++ R/rsurv-package.R
@@ -1,4 +1,4 @@
-#' The 'rsurv' package.
+#' The 'rsurv' package
 #'
 #' @description Random generation of survival data based on different survival regression models available in the literature, including Accelerated Failure Time (AFT) model, Proportional Hazard (PH) model, Proportional Odds (PO) model and the Yang & Prentice (YP) model.
 #'
@@ -6,6 +6,8 @@
 #' @name rsurv-package
 #' @aliases rsurv
 #' @import dplyr
+#' @importFrom bellreg bell
+#' @importFrom dplyr %>%
 #' @importFrom Rdpack insert_ref
 #'
 #' @references